requirements.txt
streamlit
pandas
openpyxl
plotly
seaborn
pyyaml
streamlit-authenticator
